pyproject.toml
[project]
name = "ev-grid-oracle"
version = "0.0.1"
description = "EV Grid Oracle (OpenEnv Hackathon + BESCOM Theme 9)"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "networkx>=3.2",
  "openenv-core>=0.2.3",
  "fastapi>=0.110",
  "uvicorn>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
demo = [
  "gradio>=4.44",
  "pillow>=10.3",
  "matplotlib>=3.8",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ev_grid_oracle", "server"]

